Refactor: Simplify #ifdefs for FAKEROOT_FAKENET
By splitting `communicate.h` into `communicate-sysv.c` and
`communicate-tcp.c`.

This makes it easier to see which parts are backend-specific and which
agnostic.
